add a custom database exception to the data layer.
